don't NPE if the response isn't available
We are seeing a few alerts where we hit the timeout calling Mandrill and the code tries to consume the response despite it being null.
This manifests as failing the message and us replaying, often causing duplicate emails to be sent.

Fixing this could inadvertently cause us to not send a message if the email sending times out and eventually fails.

Mandrill doesn't provide a deduplication mechanism that would nullify the concern of double sends.